Added plot view
